+++ packages/schemas/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ecoflow-api/schemas
 
+## 0.0.14
+
+### Patch Changes
+
+- c61ebea: feat: add "setEngeryManagement" function for delta 2
+- f08608d: update dependencies
+
 ## 0.0.13
 
 ### Patch Changes